Reuse markdown instance for preferences widget
